+++ packages/orq-rc/RELEASES.md
@@ -4679,3 +4679,13 @@ Based on:
 - [typescript v3.14.0-rc.68] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.68] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.68 - packages/orq-rc
+
+## 2025-10-29 09:21:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.642.1 (2.731.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.69] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.69] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.69 - packages/orq-rc